pyproject.toml
[tool.poetry]
name = "django-shop"
version = "0.1.0"
description = ""
authors = ["navruzoffd <navruzovdaniyal04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^69.5.1"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
python-dotenv = "^1.0.1"
psycopg-binary = "^3.1.19"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
